QC/requirements.txt
# 生物序列与结构分析项目依赖包
# 作者: Kuroneko
# 日期: 2025.10.04
#
# 安装方法:
#   pip install -r requirements.txt
#
# 或使用conda:
#   conda install --file requirements.txt

# 核心依赖
biopython>=1.79
requests>=2.27.0
numpy>=1.21.0

# 可选依赖（用于特定功能）
# pandas>=1.3.0  # 如需要进行数据分析
# matplotlib>=3.5.0  # 如需要绘图
# scipy>=1.7.0  # 如需要科学计算

# 开发依赖（可选）
# pytest>=7.0.0  # 单元测试
# black>=22.0.0  # 代码格式化
# flake8>=4.0.0  # 代码检查

